+++ ledger/ledger-api-tests/suites/BUILD.bazel
@@ -14,24 +14,38 @@ versions = {
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_8.scala",
         ],
     },
+    "1.14": {
+        "srcs": glob([
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/**/*",
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_14/**/*",
+        ]),
+        "entrypoints": [
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_8.scala",
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_14.scala",
+        ],
+    },
     "1.15": {
         "srcs": glob([
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/**/*",
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_14/**/*",
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_15/**/*",
         ]),
         "entrypoints": [
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_8.scala",
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_14.scala",
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_15.scala",
         ],
     },
     "1.dev": {
         "srcs": glob([
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/**/*",
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_14/**/*",
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_15/**/*",
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_dev/**/*",
         ]),
         "entrypoints": [
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_8.scala",
+            "src/main/scala/com/daml/ledger/api/testtool/suites/v1_14.scala",
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_15.scala",
             "src/main/scala/com/daml/ledger/api/testtool/suites/v1_dev.scala",
         ],
