+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_15.scala
@@ -6,10 +6,10 @@ package com.daml.ledger.api.testtool.suites
 import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
 import com.daml.ledger.api.tls.TlsConfiguration
 
-package object v1_dev {
+package object v1_15 {
   def default(timeoutScaleFactor: Double): Vector[LedgerTestSuite] =
-    v1_15.default(timeoutScaleFactor)
+    v1_14.default(timeoutScaleFactor) :+ new InterfaceIT
 
   def optional(tlsConfig: Option[TlsConfiguration]): Vector[LedgerTestSuite] =
-    v1_15.optional(tlsConfig)
+    v1_14.optional(tlsConfig)
 }
